+++ src/components/PillsContainer.tsx
@@ -1,7 +1,7 @@
 import { motion } from "framer-motion";
 import styled from "styled-components";
-import { pillVariants } from "./constants";
-import { Item } from "./types";
+import { pillVariants } from "../constants";
+import { Item } from "../types";
 
 const Base = styled.div`
   display: flex;
